feat: dns resolve in tun mode
